pass application name to release app
This change is necessary because when an app is currently deployed in
Jenkins using this repo, only the repo name of the deployed app is used
to identify the app in the release app (which keeps track which version
of an app has been released in a particular environment). There is a new
scenario where a repo is used to deploy many apps, for e.g. in the case
of licensify where 3 apps: licensify, licensify-admin and licensify-feed
are located in the same repo. So in order to identify an app, we need to
use the name also which is unique in the release app. Hence, we
introduce a new option to identify an app in the release app using the
name of the app rather than repo only. We keep backward compatibility by
still using the repo name to identify an app if the new option is not
present, this is because the name of the app may be different in this
repo and the release app.